Muevo el CustomRemoteCKAN a su propio módulo
