+++ pydatajson/ckan_utils.py
@@ -7,11 +7,9 @@ import re
 import logging
 from datetime import time
 
-from ckanapi import RemoteCKAN
 from dateutil import parser, tz
 from .helpers import title_to_name
 from . import custom_exceptions as ce
-from .constants import REQUESTS_TIMEOUT
 
 logger = logging.getLogger('pydatajson')
 
@@ -170,21 +168,3 @@ def map_theme_to_group(theme):
         "title": theme.get('label'),
         "description": theme.get('description'),
     }
-
-
-class CustomRemoteCKAN(RemoteCKAN):
-
-    def __init__(self, address, apikey=None, user_agent=None, get_only=False,
-                 verify_ssl=False, requests_timeout=REQUESTS_TIMEOUT):
-        self.verify_ssl = verify_ssl
-        self.requests_timeout = requests_timeout
-        super(CustomRemoteCKAN, self).__init__(address, apikey,
-                                               user_agent, get_only)
-
-    def call_action(self, action, data_dict=None, context=None, apikey=None,
-                    files=None, requests_kwargs=None):
-        requests_kwargs = requests_kwargs or {}
-        requests_kwargs.setdefault('verify', self.verify_ssl)
-        requests_kwargs.setdefault('timeout', self.requests_timeout)
-        return super(CustomRemoteCKAN, self).call_action(
-            action, data_dict, context, apikey, files, requests_kwargs)
